fai_inference_tp01/fai_inf01_tp01_01/requirements.txt
langchain
langchain-community
python-dotenv
langchain_fireworks
chromadb==0.4.14
langchain_experimental
pydantic
pdfplumber
nltk 
langchain_core
fastapi
uvicorn[standard]
streamlit
requests